feat(kb): editor chrome refactor + unified publish cluster
Splits the KB editor chrome into header/tab-panel/articles-actions, replaces
the legacy sidebar with a tab-driven layout, and consolidates publish controls
into a single segmented `[● Live | Publish changes · N | ▾]` cluster whose
chevron menu owns the pending list, discard, visibility toggle, and unpublish.
Also unifies KB loading states to LoadingSpinner, aligns the layout background
with MainPage, and renames the old ButtonGroup component to SegmentedControl
while reintroducing a simpler ButtonGroup primitive.
