+++ apps/web/src/app/(protected)/app/kb/layout.tsx
@@ -3,7 +3,11 @@ import type React from 'react'
 type Props = { children: React.ReactNode }
 
 function KBLayout({ children }: Props) {
-  return <div className='flex flex-1 grow overflow-hidden bg-secondary'>{children}</div>
+  return (
+    <div className='flex flex-1 grow overflow-hidden bg-neutral-100 dark:bg-background'>
+      {children}
+    </div>
+  )
 }
 
 export default KBLayout
